parsing-service/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
python-dotenv==1.0.1
httpx==0.25.2
PyMuPDF==1.23.8  # For PDF parsing
python-docx==1.1.0  # For DOCX parsing
python-pptx==0.6.22  # For PPTX parsing
pandas==2.2.0  # For data normalization
loguru==0.7.2  # For better logging
prometheus-client==0.19.0  # For metrics
pytest==8.0.0  # For testing
pytest-asyncio==0.23.5  # For async tests 